// backend/template/src/app/forum/api/model/postDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ImageEntity } from './imageEntity';
import { VoteDTO } from './voteDTO';
import { CommentDTO } from './commentDTO';


export interface PostDTO { 
    id?: number;
    content?: string;
    userId?: number;
    comments?: Array<CommentDTO>;
    votes?: Array<VoteDTO>;
    image?: ImageEntity;
    sentiment?: PostDTO.SentimentEnum;
    date?: string;
    title?: string;
}
export namespace PostDTO {
    export type SentimentEnum = 'POSITIVE' | 'NEUTRAL' | 'NEGATIVE';
    export const SentimentEnum = {
        Positive: 'POSITIVE' as SentimentEnum,
        Neutral: 'NEUTRAL' as SentimentEnum,
        Negative: 'NEGATIVE' as SentimentEnum
    };
}